Add symbol analysis to "Analyze code" functions.
Unfortunately, whether or not this will work properly is entirely
dependant on the threading subsystem and your system state. The problem
can be best described with an example sequence of events:
  - ProgramFlowAnalysis (PFA) job added to runqueue
  - SymbolAnalysis (SA) job added to runqueue
  - PFA job runs, adding more PFA jobs to runqueue
  - SA job runs, and gets none of the benfits of subsequent PFA jobs
This could be recified by adding priorities to the runqueue, or wrapping
jobs in run contexts which must be sequenced properly.

For now, you can just hit 'C' again to get the auto generated symbols.